+++ src/Command/CreateAdminUserCommand.php
@@ -34,7 +34,7 @@ class CreateAdminUserCommand extends Command
         $u->email = $this->io->askQuestion(new Question('Email (must be unique)'));
         $u->plainPassword = $this->io->askHidden('Password');
 
-        $this->userManager->update($u);
+        $this->userManager->save($u);
 
         $this->io->success('User created');