+++ src/Exception/ResetPasswordException.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Umbrella\AdminBundle\Exception;
+
+final class ResetPasswordException extends \Exception
+{
+    public string $reason;
+
+    public function __construct(string $reason)
+    {
+        $this->reason = $reason;
+        parent::__construct();
+    }
+
+    public function getReason(): string
+    {
+        return $this->reason;
+    }
+}